+++ LTU.SearchEngine.Frontend/src/Services/SearchApiFetch.ts
